requirement.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.4
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.10
marshmallow==3.6.1
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
